move all post process task to SubmissionContext
